Bug fix
HOPEFULLY, this will fix our dreaded browser hang. I can't make any guarantees this fixed it.

Added cloudflare-beacon for transparency reasons, it is not added in the site itself. It is injected by Cloudflare at a domain-browser level. I personally like to build sites that are squeaky clean, even if they are chaotic and disorganized. They at least do not track you around the web, and respect you and your hardware and your right to privacy by not shoving ads, cookies, and other useless junk in your face.